Broke out PosView tests. Added Equals test. Found bug. Fixed.
